requirements.txt
gradio
python-dotenv
prodiapy
